[refact][feat] Add FetchStateIndex for building submap
Previously, the frames used to build a submap were defined by a range of frame
indices [index + startRange, index + endRange]. Since LogStates does not always
store all of the frames, a submap could have less frames than another.
In addition, using number of frames to indicate sub map size is not intuitive.

* To build a submap around a frame, use a trajectory length criterions
  instead of using number of neighbor frames to define the submap size.

* Add function FetchStateIndex. Given a state (startId) and a trajectory length (minDistance),
  return a state iterator that its distance from startId is minDistance.

  More generally, this function can be used to fetch the state index which has
  a minDistance from a  specific state.

+++ paraview_wrapping/Plugin/vtkLidarSlam/vtkSlam.h
@@ -236,17 +236,17 @@ public:
   vtkCustomGetMacro(LoopRevisitedIdx, unsigned int)
   virtual void SetLoopRevisitedIdx(unsigned int loopClosureRevisitedIdx);
 
-  vtkCustomGetMacro(LoopQueryMapStartRange, int)
-  vtkCustomSetMacro(LoopQueryMapStartRange, int)
+  vtkCustomGetMacro(LoopQueryMapStartRange, double)
+  vtkCustomSetMacro(LoopQueryMapStartRange, double)
 
-  vtkCustomGetMacro(LoopQueryMapEndRange, int)
-  vtkCustomSetMacro(LoopQueryMapEndRange, int)
+  vtkCustomGetMacro(LoopQueryMapEndRange, double)
+  vtkCustomSetMacro(LoopQueryMapEndRange, double)
 
-  vtkCustomGetMacro(LoopRevisitedMapStartRange, int)
-  vtkCustomSetMacro(LoopRevisitedMapStartRange, int)
+  vtkCustomGetMacro(LoopRevisitedMapStartRange, double)
+  vtkCustomSetMacro(LoopRevisitedMapStartRange, double)
 
-  vtkCustomGetMacro(LoopRevisitedMapEndRange, int)
-  vtkCustomSetMacro(LoopRevisitedMapEndRange, int)
+  vtkCustomGetMacro(LoopRevisitedMapEndRange, double)
+  vtkCustomSetMacro(LoopRevisitedMapEndRange, double)
 
   // Get/Set Loop closure registration parameters
   vtkCustomGetMacro(LoopEnableOffset, bool)
